+++ Dockerfile
@@ -42,4 +42,4 @@ EXPOSE 8501
 
 HEALTHCHECK CMD curl --fail http://localhost:8501/_stcore/health
 
-ENTRYPOINT ["poetry", "run", "streamlit", "run", "BudgetChatbot_AssistAPI.py", "--server.port=8501", "--server.address=0.0.0.0"]
+ENTRYPOINT ["poetry", "run", "streamlit", "run", "Chatbot.py", "--server.port=8501", "--server.address=0.0.0.0"]
